+++ packages/langium/src/lsp/reference-finder.ts
@@ -7,10 +7,10 @@
 import { CancellationToken, Location,  ReferenceParams } from 'vscode-languageserver';
 import { NameProvider } from '../references/naming';
 import { References } from '../references/references';
-import { AstNode, CstNode, LeafCstNode } from '../syntax-tree';
+import { AstNode, LeafCstNode } from '../syntax-tree';
 import { LangiumServices } from '../services';
 import { isReference } from '../utils/ast-util';
-import { findLeafNodeAtOffset, flattenCst } from '../utils/cst-util';
+import { findLeafNodeAtOffset } from '../utils/cst-util';
 import { MaybePromise } from '../utils/promise-util';
 import { LangiumDocument } from '../workspace/documents';
 
@@ -56,7 +56,8 @@ export class DefaultReferenceFinder implements ReferenceFinder {
         const refs: Location[] = [];
         const targetAstNode = this.references.findDeclaration(selectedNode)?.element;
         if (targetAstNode) {
-            this.references.findReferences(targetAstNode, false, params.context.includeDeclaration).forEach(reference => {
+            const options = {includeDeclaration: params.context.includeDeclaration};
+            this.references.findReferences(targetAstNode, options).forEach(reference => {
                 if (isReference(reference)) {
                     refs.push(Location.create(document.uri.toString(), reference.$refNode.range));
                 } else {
@@ -65,18 +66,5 @@ export class DefaultReferenceFinder implements ReferenceFinder {
             });
         }
         return refs;
-    }
-
-    protected findNameNode(node: AstNode, name: string): CstNode | undefined {
-        const nameNode = this.nameProvider.getNameNode(node);
-        if (nameNode)
-            return nameNode;
-        if (node.$cstNode) {
-            // try find first leaf with name as text
-            const leafNode = flattenCst(node.$cstNode).find((n) => n.text === name);
-            if (leafNode)
-                return leafNode;
-        }
-        return node.$cstNode;
     }
 }
